requirements-build.txt
# Desktop packaging (macOS). Install alongside requirements.txt when building .app.
pyinstaller>=6.0.0
pywebview>=5.0.0
